chore: move ReadyToApply and ApplicantsSidebar components to components/UI
