+++ src/components/UI/common/ApplicantsSidebar.tsx
@@ -1,21 +1,20 @@
 import { Box, Flex, Link } from '@chakra-ui/react';
-import { FC, useState } from 'react';
+import { FC } from 'react';
 import NextLink from 'next/link';
 
-import { PageText } from './UI';
+import { PageText } from '..';
 
-import { SidebarLink } from '../types';
+import { SidebarLink } from '../../../types';
 
 interface Props {
   sidebarLinks: SidebarLink[];
+  sectionsInView: boolean[];
 }
 
-export const ApplicantsSidebar: FC<Props> = ({ sidebarLinks }) => {
-  const [clicked, setClicked] = useState('');
-
-  const handleClicked = (link: string) => setClicked(link);
-  // set the first link from the sidebar as acive by default
-  const isActiveLink = (idx: number, href: string) => (!clicked && idx === 0) || clicked === href;
+export const ApplicantsSidebar: FC<Props> = ({ sidebarLinks, sectionsInView }) => {
+  // mark a sidebar link as active if previous one is not visible
+  // (by the iIntersection Observer) only, to avoid having more than 1 active link
+  const isActiveLink = (idx: number) => sectionsInView[idx] && !sectionsInView[idx - 1];
 
   return (
     <Flex
@@ -34,11 +33,11 @@ export const ApplicantsSidebar: FC<Props> = ({ sidebarLinks }) => {
         <Flex key={text} alignItems='center' mb={1}>
           <Box
             borderLeft='5px solid'
-            borderLeftColor={isActiveLink(idx, href) ? 'brand.accent' : 'transparent'}
+            borderLeftColor={isActiveLink(idx) ? 'brand.accent' : 'transparent'}
             h='18px'
           />
 
-          <Box pl={3} onClick={() => handleClicked(href)}>
+          <Box pl={3}>
             <NextLink href={href}>
               <Link href={href} _hover={{ textDecoration: 'none' }}>
                 <PageText fontWeight={400} lineHeight='28px' color='brand.orange.100'>
