+++ src/components/UI/common/index.ts
@@ -1,3 +1,4 @@
+export * from './ApplicantsSidebar';
 export * from './ApplicationAttentionMsg';
 export * from './ApplicationTypeCard';
 export * from './Description';
@@ -5,4 +6,5 @@ export * from './DropdownIndicator';
 export * from './Footer';
 export * from './FooterBackgroundImage';
 export * from './ProcessStep';
+export * from './ReadyToApply';
 export * from './ThankYouBody';
