+++ src/components/index.ts
@@ -1,4 +1,2 @@
 export * from './ButtonLink';
 export * from './Nav';
-export * from './ReadyToApply';
-export * from './ApplicantsSidebar';
